+++ pkg/importer/file_test.go
@@ -1,4 +1,4 @@
-package util
+package importer
 
 import (
 	"bytes"
@@ -12,52 +12,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-const (
-	pattern       = "^[a-zA-Z0-9]+$"
-	TestImagesDir = "../../tests/images"
-)
-
-var (
-	fileDir, _ = filepath.Abs(TestImagesDir)
-)
-
 var _ = Describe("All tests", func() {
-	var _ = Describe("Copy files", func() {
-		var destTmp string
-		var err error
-
-		BeforeEach(func() {
-			destTmp, err = os.MkdirTemp("", "dest")
-			Expect(err).NotTo(HaveOccurred())
-		})
-
-		AfterEach(func() {
-			err = os.RemoveAll(destTmp)
-			Expect(err).NotTo(HaveOccurred())
-			os.Remove("test.txt")
-		})
-
-		It("Should copy file from source to dest, with valid source and dest", func() {
-			err = CopyFile(filepath.Join(TestImagesDir, "content.tar"), filepath.Join(destTmp, "target.tar"))
-			Expect(err).ToNot(HaveOccurred())
-			sourceMd5, err := Md5sum(filepath.Join(TestImagesDir, "content.tar"))
-			Expect(err).ToNot(HaveOccurred())
-			targetMd5, err := Md5sum(filepath.Join(destTmp, "target.tar"))
-			Expect(err).ToNot(HaveOccurred())
-			Expect(sourceMd5).Should(Equal(targetMd5))
-		})
-
-		It("Should not copy file from source to dest, with invalid source", func() {
-			err = CopyFile(filepath.Join(TestImagesDir, "content.tar22"), filepath.Join(destTmp, "target.tar"))
-			Expect(err).To(HaveOccurred())
-		})
-
-		It("Should not copy file from source to dest, with invalid target", func() {
-			err = CopyFile(filepath.Join(TestImagesDir, "content.tar"), filepath.Join("/invalidpath", "target.tar"))
-			Expect(err).To(HaveOccurred())
-		})
-	})
-
 	var _ = Describe("Zero out ranges in files", func() {
 		var testFile *os.File
 		var testData []byte
