+++ pkg/importer/gcs-datasource.go
@@ -15,7 +15,6 @@ import (
 	"k8s.io/klog/v2"
 
 	"kubevirt.io/containerized-data-importer/pkg/common"
-	"kubevirt.io/containerized-data-importer/pkg/util"
 )
 
 const (
@@ -117,7 +116,7 @@ func (sd *GCSDataSource) Transfer(path string) (ProcessingPhase, error) {
 		return ProcessingPhaseError, err
 	}
 
-	size, _ := util.GetAvailableSpace(path)
+	size, _ := GetAvailableSpace(path)
 
 	if size <= int64(0) {
 		//Path provided is invalid.
