+++ pkg/util/BUILD.bazel
@@ -2,17 +2,13 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "file.go",
-        "util.go",
-    ],
+    srcs = ["util.go"],
     importpath = "kubevirt.io/containerized-data-importer/pkg/util",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/common:go_default_library",
         "//staging/src/kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -23,7 +19,6 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "file_test.go",
         "util_suite_test.go",
         "util_test.go",
     ],
